Time Scroller Issue
Upon initialization the height of the time scroller container is zero. So the margin will not be set successfully when using the picker to be displayed inline! So timing out to 100 millisecond the container will be filled and it will take its actual height and there will be no problems with time scroller when using "inline" option