Fixed: today button dbl click didn't validate date
The today (home) button show the current month. After the double click it selected the current date and closed the popup, not validating the current date against the specified minimum and maximum dates.
Now, if the current date is out of the allowed range, the popup stays visible, and the input value is not changed.