Allow maxDate in past or minDate in future
Beforehand the maxDate was only allowed to be the current date (+1970/01/01) or a date in the future (e.g. +1970/01/10). You couldn't set the maxDate to yesterday (e.g. -1970/01/02) or earlier in a logical manner.

All the functionality is in there, just needed the regex opened up to check for a plus or a minus at the start of the maxDate and minDate variables and parse them if present.