Fix caret position when lazyInit and mask is set to true.
Currently when lazyInit and mask is set to true, caret position will be set at the end. This makes it unable to type the date unless we move the cursor back to the start. 

I am not exactly sure if this is the correct place to do it but it works as expected.