Today button not working anymore
Fixed a problem with the today button which is not working anymore. As
far as I could figure out the reason is
"XDSoft_datetime.prototype.setCurrentTime" which is now setting the
current time to null which doesn't reset the displayed date. Also double
clicking on the today button wasn't working because the "currentDate"
variable is null.